split code into two stores
