+++ android/src/main/java/com/dooboolab/RNIap/DoobooUtils.java
@@ -2,7 +2,7 @@ package com.dooboolab.RNIap;
 
 import android.util.Log;
 
-import com.android.billingclient.api.BillingClient;
+
 import android.content.Context;
 import android.content.pm.PackageManager;
 import com.facebook.react.bridge.ObjectAlreadyConsumedException;
@@ -47,63 +47,6 @@ public class DoobooUtils {
 
   public static DoobooUtils getInstance() {
     return instance;
-  }
-
-  public void rejectPromiseWithBillingError(final Promise promise, final int responseCode) {
-    String[] errorData = getBillingResponseData(responseCode);
-    promise.reject(errorData[0], errorData[1]);
-  }
-
-  public String[] getBillingResponseData(final int responseCode) {
-    String[] errorData = new String[2];
-    switch (responseCode) {
-      case BillingClient.BillingResponseCode.FEATURE_NOT_SUPPORTED:
-        errorData[0] = E_SERVICE_ERROR;
-        errorData[1] = "This feature is not available on your device.";
-        break;
-      case BillingClient.BillingResponseCode.SERVICE_DISCONNECTED:
-        errorData[0] = E_NETWORK_ERROR;
-        errorData[1] = "The service is disconnected (check your internet connection.)";
-        break;
-      case BillingClient.BillingResponseCode.OK:
-        errorData[0] = "OK";
-        errorData[1] = "";
-        break;
-      case BillingClient.BillingResponseCode.USER_CANCELED:
-        errorData[0] = E_USER_CANCELLED;
-        errorData[1] = "Payment is Cancelled.";
-        break;
-      case BillingClient.BillingResponseCode.SERVICE_UNAVAILABLE:
-        errorData[0] = E_SERVICE_ERROR;
-        errorData[1] = "The service is unreachable. This may be your internet connection, or the Play Store may be down.";
-        break;
-      case BillingClient.BillingResponseCode.BILLING_UNAVAILABLE:
-        errorData[0] = E_SERVICE_ERROR;
-        errorData[1] = "Billing is unavailable. This may be a problem with your device, or the Play Store may be down.";
-        break;
-      case BillingClient.BillingResponseCode.ITEM_UNAVAILABLE:
-        errorData[0] = E_ITEM_UNAVAILABLE;
-        errorData[1] = "That item is unavailable.";
-        break;
-      case BillingClient.BillingResponseCode.DEVELOPER_ERROR:
-        errorData[0] = E_DEVELOPER_ERROR;
-        errorData[1] = "Google is indicating that we have some issue connecting to payment.";
-        break;
-      case BillingClient.BillingResponseCode.ERROR:
-        errorData[0] = E_UNKNOWN;
-        errorData[1] = "An unknown or unexpected error has occured. Please try again later.";
-        break;
-      case BillingClient.BillingResponseCode.ITEM_ALREADY_OWNED:
-        errorData[0] = E_ALREADY_OWNED;
-        errorData[1] = "You already own this item.";
-        break;
-      default:
-        errorData[0] = E_UNKNOWN;
-        errorData[1] = "Purchase failed with code: " + responseCode;
-        break;
-    }
-    Log.e(TAG, "Error Code : " + responseCode);
-    return errorData;
   }
 
   public void addPromiseForKey(final String key, final Promise promise) {
@@ -151,19 +94,7 @@ public class DoobooUtils {
     }
   }
 
-  public void rejectPromisesWithBillingError(final String key, final int responseCode) {
-    try {
-      if (promises.containsKey(key)) {
-        ArrayList<Promise> list = promises.get(key);
-        for (Promise promise : list) {
-          rejectPromiseWithBillingError(promise, responseCode);
-        }
-        promises.remove(key);
-      }
-    } catch (ObjectAlreadyConsumedException oce) {
-      Log.e(TAG, oce.getMessage());
-    }
-  }
+
 
   public WritableMap convertJsonToMap(JSONObject jsonObject) throws JSONException {
     WritableMap map = new WritableNativeMap();
