+++ android/src/play/java/com/dooboolab/RNIap/RNIapPackage.java
@@ -24,7 +24,6 @@ public class RNIapPackage implements ReactPackage {
   public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
     List<NativeModule> modules = new ArrayList<>();
     modules.add(new RNIapModule(reactContext));
-    modules.add(new RNIapAmazonModule(reactContext));
     return modules;
   }
 }
